requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
decorator==5.1.1
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.5.0
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
rest_framework_simplejwt
redis
celery
django-celery-beat
flake8
flake8-isort
flake8-html